checking agains divided by zero, and cleaner round up with BigDecimal and RoduningMOde

+++ src/main/java/de/papenhagen/MeasuringResource.java
@@ -33,7 +33,7 @@ public class MeasuringResource {
                 .getCurrentMeasurement()
                 .getValue();
 
-        double convertLvl = convertMeasuringUnit.convert(levelOfSanktArnual);
+        int convertLvl = convertMeasuringUnit.convert(levelOfSanktArnual);
 
         return new BottleCount(convertLvl);
     }
