+++ src/main/java/de/papenhagen/entities/BottleCount.java
@@ -11,5 +11,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class BottleCount {
-    double count;
+    int count;
 }
